requirements.txt
celery==4.2.1
coverage==4.5.3
django>=1.11.19
django-cors-headers==3.0.2
django-extensions==2.2.1
django-localflavor==2.2
django-phonenumber-field==3.0.1
djangorestframework==3.7.7
djangorestframework-jwt==1.11
phonenumbers==8.10.15
pillow==6.1.0
psycopg2==2.7.3.2
pytest-cov==2.7.1
pytest-django==3.5.1
redis==3.2.1
requests==2.16.0

-e git+ssh://git@github.com/baronvonvaderham/django-mtg-card-catalog@master#egg=django-mtg-card-catalog
